src/preprocessing_question_gen/requirements.txt
PyPDF2==3.0.1
python-docx==0.8.11
textract==1.6.3
google-cloud-storage
streamlit
numpy